Demostrative parallel execution using ray
Jobs are run in parallel in a local ray cluster.

It's just a quick demo to investigate if this is a path worth
pursuing.

There is a new LogConfig class used to handle log configuration, this is
a good idea overall but was needed to pass log configuration to ray
tasks in a reasonable way.

One can specify the same job multiple times in the pipeline steps if
a name is also provided for that step.

I did this very quickly and there are some minor caveats:
- Job hooks do not work
- ray prints the pid of the running process when logging to console but
  the logfile doesn't contain any info on the task (so neither on the
  order of the lines in the file, all jobs are just printing
  concurrently and that's definetly not good)